Implicit library declaration and refactor code
