+++ src/main.c
@@ -6,56 +6,16 @@
 /*   By: ribana-b <ribana-b@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/15 00:25:02 by ribana-b          #+#    #+# Malaga      */
-/*   Updated: 2024/03/20 11:56:44 by ribana-b         ###   ########.com      */
+/*   Updated: 2024/03/21 02:38:01 by ribana-b         ###   ########.com      */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-int	stack_len(t_stack *stack)
-{
-	t_stack	*temp;
-	int		counter;
-
-	counter = 0;
-	if (!stack)
-		return (0);
-	temp = stack;
-	while (temp)
-	{
-		++counter;
-		temp = temp->next;
-	}
-	return (counter);
-}
-
-void	update_index(t_info *info)
-{
-	size_t	index;
-	t_stack	*temp;
-
-	index = 0;
-	temp = info->stack[A];
-	while (temp)
-	{
-		temp->index = index++;
-		temp = temp->next;
-	}
-	index = 0;
-	temp = info->stack[B];
-	while (temp)
-	{
-		temp->index = index++;
-		temp = temp->next;
-	}
-}
-
 static int	push_swap(t_stack **stack_a)
 {
 	t_info	info;
 
-	if (!(*stack_a))
-		return (1);
 	initialise_info(&info, stack_a);
 	sort_number(&info);
 	*stack_a = info.stack[A];
